sql/dogs/01_initial_draft/requirements.txt
mysql-connector-python
python-dotenv
pycodestyle